Move remaining dummy drivers to servers.
