+++ drivers/register_driver_types.cpp
@@ -30,7 +30,6 @@
 
 #include "register_driver_types.h"
 
-#include "drivers/dummy/display_server_headless.h"
 #include "drivers/png/image_loader_png.h"
 #include "drivers/png/resource_saver_png.h"
 
@@ -38,8 +37,6 @@ static ImageLoaderPNG *image_loader_png;
 static Ref<ResourceSaverPNG> resource_saver_png;
 
 void register_core_driver_types() {
-	DisplayServerHeadless::register_headless_driver();
-
 	image_loader_png = memnew(ImageLoaderPNG);
 	ImageLoader::add_image_format_loader(image_loader_png);
